# Conflicts:
#	SearchLite.sln

+++ Source/SearchLite/Interfaces.cs
@@ -1,17 +1,20 @@
 ﻿namespace SearchLite;
 
 /// <summary>
-/// Common interface for searchable documents
+/// Common interface for searchable documents.
 /// </summary>
 public interface ISearchableDocument
 {
     /// <summary>
-    /// Document key WITHIN the collection. Deletes and updates will be based on this key. This is not a global unique identifier, but must be unique within the collection.
+    /// Document key within the collection. Deletes and updates are based on this key. It is not a
+    /// global unique identifier, but must be unique within the collection.
     /// </summary>
     string Id { get; }
+
     /// <summary>
     /// The text that will be included as input to the full text search engine.
-    /// Other parts of the document will be ignored for the purposes of full text search, but can be used for filtering and ordering.
+    /// Other parts of the document are ignored for full text search, but can still be used for
+    /// filtering and ordering.
     /// </summary>
     string GetSearchText();
 }
@@ -28,7 +31,7 @@ public class SearchResult<T> where T: ISearchableDocument
 public class SearchResponse<T> where T: ISearchableDocument
 {
     public required IReadOnlyList<SearchResult<T>> Results { get; init; }
-    public int TotalCount { get; init; }
+    public long TotalCount { get; init; }
     public float MaxScore { get; init; }
     public TimeSpan SearchTime { get; init; }
 }
